glibc_multi: fix invalid 32bit glibc when cross compiling
The issue this fixes happens when cross compilation is used. The
pkgsCross.*.buildPackages.glibc_multi actually combined x86_64 with
target's. That is because pkgsi686Linux actually is package set where
target and host are the same and thus buildPackages.pkgsi686Linux
actually is the same as using directly pkgsi686Linux. To overcome that
we need packages that have cross platform target but host and build
platform has to be set to i686.

This change does not affect usage outside cross compilation because in
that case buildPackages set is the same as the current set.