python: 2.7.18.6 -> 3.10.10
This changes the top-level attribute `python` to refer to `python3`
instead of `python2`.

This allows users to get a supported version of one of the most popular
programming languages if they run `nix shell nixpkgs#python` et al.

As `python2` is already marked insecure, older repos built with a newer
nixpkgs and referring to `python` would currently throw an error.
With this change, they would silently switch to `python3` if the
changelog entry isn't read.

An alternative would be to remove the alias and to require users to
explicitly refer to either `python2` or `python3`.