stdenv: add buildCanExecuteHost
This PR adds `stdenv.buildCanExecuteHost`, which is defined as:

  buildCanExecuteHost = stdenv.buildPlatform.canExecute stdenv.hostPlatform;

By making the above expression referenceable by the nice short
string `stdenv.buildCanExecuteHost`, I hope people will find it more
convenient to type and therefore be more likely to use it instead of
`stdenv.buildPlatform==stdenv.hostPlatform`

I think that `stdenv.buildPlatform==stdenv.hostPlatform` is almost
always the wrong thing, and we should start discouraging it.  This
PR adds stdenv.buildCanExecuteHost, which is almost always what
should be used instead.

My buildfarm is a bunch of AMD Opterons.  I recently switched over
to expressing the `-march=` flag for them the correct way: by
creating a new local `buildPlatform` called `opteron` for them which
sets `gcc.arch="bdver1"` (and equivalent for `rustc`).

I have a few other `x86_64` machines which do not support the
`bdver1` instructions (mostly the AVX/SSE stuff).  So these machines
will segfault if you try to use an `opteron`-compiled `ffmpeg` on
them.  Therefore, I do two builds of nixpkgs for x86_64 in parallel:
one for `opteron` and one for `lib.systems.examples.gnu64` (which is
the "vanilla" `x86_64`).

Here's the problem:

```
  stdenv.buildPlatform != stdenv.hostPlatform
```

because

```
  opteron != lib.systems.examples.gnu64
```

... but this really a cross-compile from the perspective of
everything outside of `stdenv`.  Most packages that are doing an
equality-test between `buildPlatform` and `hostPlatform` really just
want to know if the build platform can execute host binaries --
usually because upstream's build machinery is broken in that case so
features have to be disabled.

Unfortunately we've established
`stdenv.buildPlatform==stdenv.hostPlatform` as the idiom for "is
this a cross compile?".  So I'd like to make it *more* convenient
for package authors to use `stdenv.buildPlatform.canExecute
stdenv.hostPlatform`, by making it less verbose for them to type.