python3.pkgs.ray: fix evaluation when generating nix dependency graph
Due to some oddities in the language `builtins.tryEval` fails to
catch some instances in certain edge cases. While it would be nice
to have them fixed in nix itself for now it seems way easier to
just remove these two instances.

This fixes evaluating nativeBuildInputs in
https://github.com/tweag/nixpkgs-graph/

Retrieve this information from nixpkgs is quite complex but having it
would enable a lot of interesting use cases.