recastnavigation: fix for darwin build break
Building on Darwin resulted in an error with not being able to find SDL header files.
Looking through the remote repository code, it looks like it expects the SDL2.framework layout
and not the Linux-style include directory layout.

The main difference between these two layouts is whether or not the SDL header files are
contained within a `SDL2` subdirectory. In the .framework file, the header files are located
at `Headers/SDL_*.h`. On Linux, the headers are located under `include/SDL2/SDL_*.h`.

When Nix builds SDL on Darwin, the output is in the Linux-style layout, while the
recastnavigation CMakeLists.txt is expecting things to be in the .framework layout.

The fix I made was to change the Darwin-specific `include_directories` directive which points
to the Headers directory to instead point to the headers from the Nix store.