llvmPackages_*.clang: bind linker on compiling clang
When using clang for cross compilation, clang cannot find the linker.
Because, for consistency with gcc, given a linker name, clang finds the
linker in the following order:

1. in prefix dirs, which contains toolchain specific paths and paths
   specified by `-B`
2. in ProgramPaths
3. in PATH environment

(refer to `Driver::GetProgramPath` in `clang/lib/Driver/Driver.cpp`)
(also refer to https://gcc.gnu.org/onlinedocs/gccint/Collect2.html)

The step 2 and 3 considers all possible target triple prefixes, but
the step 1 does not. In nixpkgs, all binaries in cross toolchains are
prefixed by the target triple, thus clang cannot find them.

For gcc in nixpkgs, the linker path is hardcoded in configuration
phase. This commit copies this behavior to clang.

But the added cmakeFlags also introduce circular dependency. The
circle comes from the fact that clangNoLibc and clangNoCompilerRT
introduce dependency of libc unexpectedly via cmakeFlags of their
underlying clang. We need to override bintools for them. Refer to

https://github.com/NixOS/nixpkgs/pull/220595#issuecomment-1787393460

for details.