python3Packages: don't propagate build-platform cffi
cross-compilation. This had the unfortunate effect of making build-platform
cffi part of the runtime closure of all downstream packages.

On the other hand, it appears that it is not actually necessary to propagate
cffi to fix cross-compilation; instead it is enough to add it to
nativeBuildInputs.

This patch implements this change, while also reverting #200779 (my initial
attempt at fixing build-platform contamination in brotlicffi) and removing a
few other propagatedNativeBuildInputs that were unnecessarily added in #142301.