nixos/lib/testing: allow overriding `virtualisation.test.nodeNumber`
Say you have a VM test with two nodes, but only one node is started and
eventually switched to the config of the other, i.e.

    nixosTest {
      name = "foobar";
      nodes = {
        node1 = {};
        node2 = {};
      };
      testScript = { nodes, ... }: ''
        node1.start()
        node1.succeed("${nodes.node2.config.system.build.toplevel}/bin/switch-to-configuration test")
      '';
    }

then you'll end up with another IP (i.e. `192.168.1.2` instead of
`192.168.1.1` on `node1` in the example above). This can be a problem
when writing a test where a node's configuration must be changed while
running and the IP must remain stable.

The workaround I ended up with was

    virtualisation.test.nodeNumber = lib.mkForce 1;

However this option is read-only, so currently this breaks with an
evaluation error.

Since I haven't found a trivial way to override
`virtualisation.test.nodeNumber` (another module
redefining the option isn't accepted[1]), I figured it's OK to drop the
immutability here.

[1] I tried to work around like this:

        let
          hackhack.options.virtualisation.test.nodeNumber = lib.mkOption {
            type = lib.types.int;
            default = 23;
          };
        in nixosTest {
          # ...
          nodes.node1.imports = [ hackhack ];
        }

    But it broke with the following error:

        error: The option `nodes.cfssl.virtualisation.test.nodeNumber' in `makeTest parameters' is already declared in `/home/ma27/Projects/mf-admin/nixpkgs/nixos/lib/testing/network.nix, via option extraBaseModules'.