buildRustCrate: Add support for build kinds
Adds proper support for build kinds. Based on cargo.

A new top-level parameter `buildKinds` can specify which build kinds should
be compiled. By default for backwards compatibility the build kinds are
["lib" "bin"], with "test" added if `buildTests` is true.

Currently available build kinds: ["lib" "bin" "test" "example" "bench"].
When a build kind is disabled, all builds of that kind will be skipped.

Additionally, all values in `crateBin` can now take new parameters. This
is to set the kind of binary should be built, and whether it should be
benched or tested. For backwards compatibility if not set, it is of kind
"bin" along with `test` set to true when `buildTests` are true.

```nix
{
  name = "xxx"; # existing
  src_path = "xxx"; # existing
  kind = "bin"; # Same values as `buildKinds` with exclusion of lib
  harness = true;
  test = true;
  bench = true;
  harness = true;
}
```

Setting harness determines whether to use the libtest harness, as in
cargo. For the library, use the top-level `libHarness` parameter.
It only matters when compiling for bench or test kinds.

As before, it is up to the user to create multiple `buildRustCrate`
derivations providing the proper dependencies for different kinds.

Autodiscovery is performed, as in cargo 2015, when no binaries of the
specified kind are set in `crateBin`.

There are no breaking changes to the buildRustCrate api. But this
deprecates the use of `buildTests` in favor of `buildKind = ["test"]`