mesa: build dri for arm
The isAarch64 applies only to 64bit arm. The commit
e7a2c65ab54f096546c9aad6ae2e4e29e61f2885 missed support for 32bit arm.

This also removes kmsro as it is not used by Mesa meson.build file. The
kmsro is detected from other galium drivers. This might not be correct
and there probably should be a way to enable this in the upstream but
for now this does nothing.