imagemagick: use the right pkg-config in *-config
If we're building buildPackages.imagemagick, the imagemagick libraries
will be built for the local system, but we were previously using the
pkg-config for the cross system in the *-config scripts bundled with
ImageMagick.  I suspect the reason it wasn't noticed for a long time
was that packages that are using still old-style *-config scripts
instead of pkg-config probably don't cross compile anyway.  I noticed
because I wondered why buildPackages.imagemagick when cross-compiling
was not the same as my native imagemagick, so as a happy side effect
of this change, users won't have to rebuild imagemagick any more when
it's used as a native build input for a cross-compiled package.