mwprocapture: 1.3.0.4236 -> 1.3.0.4328
Version V1.3.0.4236 of mwprocapture (the Linux driver for the Magewell Pro Capture family) FTBFS when building against Linux 6.1 or newer. 

This patch bumps the driver version to 1.3.0.4328 that Magewell published to address this issue. The 1.3.0.4328 release notes state:

> Fix problem: driver installation may fail on an operating system with kernel version 6.1 or 6.2.

pci.patch has also been dropped as that fix is now applied upstream.