init: lib.flattenAttrs
This change adds lib.flattenAttrs function that flattens an attribute
set. Along with lib.collect', it is mostly inteded for use with Nix
flakes where packages must be a flat attribute set of derivations. That
is, this allows flakes to define packages (or other outputs) as a tree
of derivations (or other types) and flatten them before exposing.

In particular, this is useful for flakes that want to expose
cross-compiled variants of packages so the user can run, for example,
`nix build path:.#go-linux-amd64-v3` independent of the current system
(as in `packages.${system}`). In this case, Go is used as an example
since the build system has an excellent support for cross-compilation
out-of-the-box (assuming Cgo is disabled).