xorg.libpciaccess: fix pci.ids lookups at runtime
In GitHub issue #221592 I found that virt-manager was not showing PCI info.
Tracking the issue down points to this library, which is consumed by libvirt.
libpciaccess was reading from the default path of /usr/share/hwdata/pci.ids,
which will likely not exist on NixOS.

Instead of using the hard-coded path, change the libpciaccess package to depend
on hwdata at runtime and to be configured with the hwdata share path.

Updated the generation script, as well as committed the newly generated changes.