python3Packages.mypy: fix setuptools integration
This patch updates the patch used to fix compatibility with the latest version
of types-setuptools. The previous version of the patch caused mypycify() to fail
with the following error:

UnboundLocalError: local variable "extension_class" referenced before assignment

The new version of the patch fixes this issue and is the version that was
accepted upstream.