bpftools: 5.19.12 -> 6.2; sync with linuxHeaders
bpftools used to use the same version and sources as linux_latest, but
this was changed in f12d2f016b6 ("bpftools: decouple version from
linux_latest"), because it caused too many rebuilds.

Since then, it has languished.  It came to my attention because
bpftool btf dump seems to be completely broken on recent kernels.
So, since it's a package that really needs to be kept up to date to be
compatible with the latest kernels, let's couple it to linuxHeaders
instead.  linuxHeaders updates already have to go through staging, so
we won't suffer from extra rebuilds like we did when it was coupled to
the latest kernel directly, which goes straight to master.