perl536Packages.Po4a: remove texlive dependency
The motivation behind this is wanting to reduce the reverse closure size of
texlive.

When internationalizing tex files, Po4a attempts to follow included files. In
order to find the correct files, po4a uses texlive's file-finding utility
kpsewhich. This commit replaces texlive with a tiny shell script, which is
precisely good enough to run the tests.

This does not change the resulting package in any way, it only affects how the
test is run.