tabula: pin jre version
The packaged (latest) release of `tabula` is quite old, and bundles a very
old version of `jruby` which is incapable of running on modern JREs.

See https://github.com/tabulapdf/tabula/blob/v1.2.1/.ruby-version#L1 for package version, and
https://github.com/tabulapdf/tabula/blob/master/.ruby-version#L1 for current (but unpackaged)
version.

The `jruby` bug which prevents use on modern JREs was apparently fixed in version 9.2.8.0