maintainers/team-list: add 3 new members to LLVM
The three of us have been working on getting the different LLVM
versions in sync, and discussing plans for 16, 17, and upgrading the
default.  Being notified of PRs will be useful, so that we can make
sure that new PRs don't introduce further unnecessary deltas between
different LLVM versions, which has been a problem so far.