espeak-ng: Add alsa-plugins path to ALSA_PLUGIN_DIR
Fixes https://github.com/NixOS/nixpkgs/issues/222722

On non-NixOS systems, `nixpkgs#espeak-ng` currently fails to find
`libasound_module_pcm_pulse.so`, as it seems to look in `alsa-lib-1.2.8`
but the file in question exists in alsa-plugins-1.2.7.1.

This change points it to the correct location allows `espeak-ng` to
work without error.