ghidra: Add lldb support
Using the lldb-based debugger in Ghidra requires some care. The native
interface needs to match the used lldb version, a library needs to be
built and everything required needs to be on the path.

I was able to get into an lldb debugging session with this change. I was
not able to set breakpoints or go any further. I hope this is an
improvement nonetheless.