xdg-desktop-portal-kde: fix missing runtime dependency
`plasma-workspace` is required when `xdg-desktop-portal-kde` needs to
open an application picker to let the user pick the application to use to
open a certain file (when no existing default is available).
If `plasma-workspace` is not available the `xdg-desktop-portal-kde` service
logs an error and nothing else happens (the application picker is not opened).

See https://github.com/NixOS/nixpkgs/issues/160923#issuecomment-1474989176