texlive.bin.core: fix cross compilation
The build system generates the file "hitables.c" with the program
"./himktables" that is also compiled by the build system.

This does not work when cross-compiling since this program is built for the
wrong architecture. We already use a natively built copy of texlive.bin.core for
various other tools (like tangle, weave), so this change just adds himktables to the
this native copy and uses this program in the cross build.