nodejs: fix cross compilation for armv7l
The v8 library supports only limited number of build platforms based on
the host architecture. The rule there is the bitness (the mixing of
32bit and 64bit architectures seems to be in most cases disallowed).
Thus this adds usage of the emulator of the host platform and builds
tools for that. The overall build slowdown is negligible while at the
same time it ensures

The patch for the emulator wrapper application was sourced from
Buildroot project.