xfce4-panel: Allow embedding plugins in derivation.
This change allows xfce4-panel plugins to be embedded in the xfce4-panel derivation and be subesquently found when the xfce4-panel program is launched. This allows xfce4-panel plugins to be used outside of a NixOS system configuration (i.e. non-NixOS systems, home-manger, etc.). This is a non-breaking change (NixOS users continue to specify their plugins as usual).