vimPlugins.copilot-vim: set default g:copilot_node_command
copilot-vim requires NodeJS to function properly since the plugin
bundles JS files. You can set the path to the NodeJS that should be
used manually, in your configuration.

We can set the default value to a valid NodeJS binary, so in the general
case, users don't need to do anything for this plugin to work.

Users can still set g:copilot_node_command manually, if they want to.