correct key name in output of ssh-keygen command
the ssh-keygen example specifies `example-key` as the name for the key in the command line, then uses `test-key` in the example output. That's not the output that would be generated from executing the command and it's probably a remnant of a previous version of this example. Later in the example for the filesystem configuration, example-key is used as expected.