cppzmq: make zeromq propagatedBuildInput
cppzmq is a header-only library. As such, it is not much useful
without zeromq, whose headers it includes. By having zeromq in
propagatedBuildInputs, we can simplify dependent expressions.

This change is motivated by
https://github.com/lopsided98/nix-ros-overlay/issues/255#issuecomment-1487590226.
The expressions in nix-ros-overlay are automatically generated and
since packages in other distributions where ROS runs need not to
explicitly depend on zeromq, the Nix expression should behave the
same. This way, nix-ros-overlay will not to have manually
patch/overlay the automatically generated expressions.