pkgsStatic.iptables: fix build
As a result of 8f45423217c ("libpcap: build with libnl"), the static
build started to fail, because iptables didn't know to link with
libnl.  The patch makes iptables use pkg-config to find libpcap, which
means it knows about other libraries that need to be linked when doing
a static build.