memreserver: init and add module
This adds memreserver, a small utility which implements a workaround for
the inability to reliably enter deep sleep states on systems with
dedicated GPUs (i.e. those containing their own VRAM).

Linux's current PM system which is responsible for putting hardware to
sleep in an orderly fashion currently has only hierarchical
relationships and no phases, thus when the GPU is getting suspended the
kernel might have already suspended the disks. The process of suspending
a GPU with VRAM needs significant amounts of memory as all contents from
the GPU VRAM need to be moved to system RAM as GPU VRAM will be fully
powered down and thus lose its contents when entering deep sleep.
But if the disks are already suspended the kernel can no longer swap out
additional memory nor release all types of disk caches, thus even with
plenty of nominally free memory the suspend can fail due the kernel
running out of a specific allocation type. On anything but the newest
kernels (6.1+) at least on AMD this cascades into a full resume failure
which can only be fixed by a reboot.

memreserver works around this by forcing the kernel to free enough
real system memory to fit all of the GPU VRAM before allowing the system
to proceed to suspending devices. During this phase the kernel can still
flush caches and swap out data to a swap device, thus unless the system
is completely out of system memory as well as swap this will succeed.
Afterwards when the kernel suspends the GPU it now always has enough
free system memory to evacuate the entire GPU VRAM into.

It works by listing all known GPUs, checking if they have VRAM which
needs to be evacuated, summing the amount of used VRAM and then
allocating this amount as locked memory (forcing the kernel to move
everything else out of it) and immediately freeing it, thus allowing the
kernel to use it later. It currently only works for AMD GPUs, but can be
extended if we have testers for Intel dGPUs and/or nouveau.